requirements.txt
shodan
psycopg2-binary
requests
python-dotenv
